Removing CharSet
We have both CharSetGrammar and CharSetExpr, but we don't use it in the actual sheets, it's just there for about a dozen tests.  And it's a lot of code, Grammars and Exprs are big.

So this makes `CharSet(tape, chars)` a convenience function for `Uni(chars.map(c => Lit(tape,c)))` instead.