Unifying deriv results & eliminating "open" derivs
openDeriv/openDelta were a bit annoying; only a few grammars had them, but the root had to, so everything that could contain Priority had to be given openDeriv/openDelta, and also check that its child had it as well.  That was annoying, and getting in the way of some other stuff I wanted to do.

This commit goes back to a much earlier idea, that deriv/delta can take a magic "open query" tape.  (Note that this isn't actually going back to the old implementation, though.  The old implementation made it so "__ANY_TAPE__" matched any tape, like the deriv with respect to __ANY_TAPE__ of `t1:a` was `t1:a`.  Here, it doesn't match, just like ".END" and other magic tapes.  The only thing sensitive to this special tape right now is Priority.)

(1) To accomplish this, deriv and openDeriv, and delta and openDelta, had to be refactored so that they have the same signatures.  The difference was that deriv and delta *didn't* return the tape they acted on (because it was obvious and invariant, it was the same tape they were queried on).  Now they do.

(2a) We already have a name for a [tapeName, text] pair -- it's a Literal.  That's what motivated a previous refactoring, that Outputs were no longer their own hierarchy where there's a special [tapeName, text] object put into a trie -- they just became LiteralExprs within ConcatExprs.

Now we do the same with derivative results, those are now expressions like Outputs are.  We take the result of the deriv and place it right in the Output trie.

(2b) Looking at it, though, LiteralExpr itself is a bit heaviweight, like it's got the question of atomic text vs tokens, indexing, and LTR vs. RTL.  None of that is relevant for the one-token literals we're concerned with in deriv and Outputs.  So I made a simple TokenExpr representing a one-char literal.

(2c) To represent the Epsilon token, we can just use EpsilonExpr.

(3) We were a bit all over the place with the names of input/output variables for child derivs.  While I was busy replacing tapeName/target in the param list for derivs, I rectified the names to be more consistent.

   * We used "target" for the initial parameter passed into deriv, now that's consistently "query"
   * We also used "target" for the result returned from child derivs, now that's consistently "result" (usually "cResult")
   * We usually used "next" for the successor variable returned from child derivs, but occasionally that was "expr", so now it's consistently "next" (usually "cNext").