Removing ANY_CHAR_STR in favor of DotExpr
This fixes a very silly bug, but in the end is probably for the best anyway.

(1) The reintroduction of "any tape" queries in the last commit got me thinking about what to name it.  It used to be called "__ANY_TAPE__" but we've gotten more strict about identifiers since then.  "__ANY_TAPE__" is a valid identifier, so it shouldn't be that; the proper way to do name now would be something like "$any" or "$open".

That raises the issue of what ANY_CHAR_STR should actually be.  It's long been "__ANY_CHAR__", but that's a valid literal... it's a very UNLIKELY literal, but it's valid.  And with atomicity, it's a valid token, too.   Any sequence of characters is a valid literal, and thus, with atomicity, a valid token.

Really we should be doing some other kind of test other than `query.text == "__ANY_CHAR__"`, like an instanceof test.  Now that queries and outputs are just Exprs, we alrady have an Expr that means "text is any char", that's DotExpr.  So the query is now just a DotExpr rather than a magic string.

(2) I also moved `expandStrings()` out of the Tape object.  Now that that doesn't take a magic string anymore, we don't want pass in the actual Expr to Tape (for it to check whether it's a specific char or any char by `instanceof`), because that results in a circular import and TypeScript can have trouble with those.  So now it's a method on TokenExpr and DotExpr -- it basically means "what vocab can you represent".

(3) There's a thing in MatchFromExpr where we were expanding a string on fromTape and then filtering it with toTape, regardless of which tape the result came from.  (This works because it's symmetrical, we're just getting the intersection.)  This doesn't work with the new expand strings, because we don't pass in a separate string anymore -- with TokenExpr.expandStrings() it's got to be that token's text.  So instead we expandStrings on the result and filter on the oppositeTape.

(4) This reminds me, I gave up in the last commit making all these changes to the original MatchExpr -- I couldn't quite get it to work.  (I'm sure it's possible, it's just that what MatchExpr is doing with tapes and vocabs is a lot more complex than other expressions.)  I just commented it out; we haven't used it in a long time anyway.