Adding Priority tests
We weren't testing Priority on its own, so that makes it dicey to refactor safely.

I also added some tests for behavior that's beyond the current capabilities of Priority, but that I think are worth working towards:

    * It's not currently possible to nest Priority and have them both work.

    * It's not currently possible to express Priority partially, (i.e for a subset of known tapes).

These don't just fail to get the right answer, they throw an unsimplified result exception.

Not having these behaviors is ultimately the cause of the tricky and brittle logic right before constructExpr involving Priority, Collection, and Counts.  There might or might not be a Priority, but there can only be one and it has to be complete, so you have to find it before making another, and to make that easier we expect it in certain structural positions.

If Priorities were properly nestable, this logic would become simple.  We'd not have to look for the existing Priority, just wrap the whole thing in a new Priority that handles any unhandled tapes.