Recovery of internal (e.g. PreTape) outputs
This commit is just a proof-of-concept, it doesn't yet work for the general system, but it illustrates that it's actually pretty simple to extract "hidden outputs" from deep in the tree.

Some expressions (like PreTapeExpr and the experimental HideExpr) are executing on tapes that were only valid in their scope; the top-level Priority doesn't know about them.  This way of executing/ordering tapes ends up being more robust than the relatively brittle PriorityExpr, but there was a fatal downside, that they threw out their results.

This commit introduces OutputExpr.  It's a holder for "finished" material (material that's already been delta'd/deriv'd to completion), but in most other situations acts like an Epsilon: it's nullable on all tapes, it has no deriv, and it's a valid output at the end of generation.

(OutputExpr also takes the "be the skeleton of the output trie" responsibility back from ConcatExpr.  It ended up being simpler that way.  Really, OutputExpr is the same as the old OutputTrie object, it's just that it now has a delta and deriv and can thus sit in the expression tree.)

Stepping back, we can think of an expression as being one of three types:

    (1) Most expressions we're concerned with are "unfinished" (they might have material that hasn't yet gone through the generation loop) and have "unlicensed" tapes (the expression references tapes that aren't necessarily handled by any outer expression).  Plain old `t1:hello` is one of these.

    (2) Outputs are expressions that are finished -- they have no material that hasn't gone through generation.  (Their tapes don't have to be licensed, licensing is something you need to go *through* the generation loop, and these are already through it.)  OutputExpr is one of these, obviously, but so are EpsilonExpr and NullExpr, trivially so.

    (3) There are also "tape licensors" like PriorityExpr and PreTapeExpr that license otherwise-unlicensed tapes in their scope.  An expression where all the tapes are licensed isn't necessarily finished yet, but it has all the material itself to under generation.

Anyway, this commit doesn't yet completely thread the new Output idea through the whole system.  It really only works for some toy examples so far.