Further tape-licensing proof of concept
Warning, be careful executing the whole test suite here, most will be infinite loops.

This expands on the proof of concept in the previous commit, with some examples of simple grammars using a minimal licensor "TapeExpr".

(1) It wasn't, in the end, possible to use PreTapeExpr itself as a general tape licensor.  PreTapeExpr was written in a narrow circumstance, where the two tapes have exactly the same number of outputs in PreTapeExpr's scope.  In cases where they DON'T, the parse fails -- which is appropriate in that circumstance, but not otherwise.

(2) In the general case, two tapes might have different numbers of outputs.  This means that t1 might get a delta result in the same cycle as t2 gets a deriv result, or the other way around.

This is actually awkward for us here.  It's not awkward at the top-level, like where the generator loop is calling both delta and deriv, or even when the query is generated near the top like in PriorityExpr.  But it's awkward when we're arbitrarily far down in the tree, because delta and deriv are different traversals of the tree.  We end up having to call a child.delta from deriv and a child.deriv from delta.  And we can do that (other derivs call child.delta, after all) but it's awkward, it illustrates that something is wrong.

What's wrong, really, is that the generator loop is neither delta/deriv alone, but both.  If we want to push a part of this loop potentially deep into the tree, we really need a function that's both.  (It's also the awkwardness with holding onto generated deltas in order to disjoin them with derivs, we have to do that in generate because there's no other functional home for that operation.)

So we need that function back; I've called it `forward` because it advances all of our [Output, Expr] "cursors" forward one sep.

(3) The generate loop no longer concerns itself with delta/deriv, it only calls "forward".  Forward itself has no token result and the generate loop doesn't keep track of outputs anymore, all the outputs are in the tree now.

Not having explicit outputs made it difficult to read debug logs, those were important waystones, so I expanded `toDenotation` to all Exprs, what it means now is "Give me all the Outputs you can currently reach and turn them into a StringDict".  It gives roughly, but not exactly, the same outputs as before.

(4) This is my second experiment so far with nestable tape systems, and one thing I've found is that they make it hard to catch when you've got invalid, not-simplified results.  Every tape licensor (here, `TapeExpr`) needs to fire the open query (here, `forward()`) to its child, without knowing whether that will be handled by a nested licensor later on.  If there's not a licensor, it needs to be a no-op (you don't want it to FAIL), but that means that if you apply it to expressions where there's no licensor, generation goes into an infinite loop.

We could catch that if we had equality relation on grammars, but we don't (and it'd be expensive besides).  It's probably better just to return some kind of sign, like a boolean that the open query was handled in a non-trivial manner.  I haven't done that yet, though, and haven't finished threading this system everywhere either, so beware of infinite loops literally everywhere now.

(5a) Even if there were no unlicensed tapes, we still wrapped the whole thing in a PriorityExpr, which we probably shouldn't.  This makes it so that when tapePriority.length == 0, we don't actually create one.

(5b) I caught a bug related to this, in priority calculation.  We indicated tapes that were already licensed with negative numbers, but at the end multiplied that number by their vocab size in order to prioritize high-vocab tapes.  An already-licensed tape with no vocab results in a priority of -0.  This is NOT less than zero in JavaScript, so the tape ends up being considered unlicensed again, and then the priority failsafe re-licenses it.  Now vocab sizes are considered to have a minimum size of 1 for the purposes of this calculation.