Nested cursors instead of Priority
(1) TapeGrammar/TapeExpr wasn't a very descriptive term.  They're now `Cursor`, in the sense that they represent a cursor in between the finished output expression and the unfinished remainder, and are the thing that goes `forward`.

(2) Rather than wrapping everything in `Priority()` to handle any tapes still unlicensed, we use the convenience function `Cursor()` creates a nested Cursor that does the same thing.  You can use it either with a single tape name or a list of them.

(3) Factored out the process of disjoining deriv results from `Expr.disjointDeriv` into a function `disjoin`.  (E.g. instead of calling `this.child.disjointDeriv()`, call `disjoin(this.child.deriv())`.)

This lets us disjoin lists of results that aren't exclusively from `deriv`, like when we want to disjoin the results of both `delta` and `deriv`.