Improved logging of Cursors
As I was threading Cursors and Outputs through the system, and encountering bugs, I was getting stymied by the fact that these changes had broken the expression logging system.  So this commit addresses that.

(1) There were no longer prevOutputs at the top level, but prevOutputs were the main signposts we used to understand the deriv algorithm (e.g. what expression was derivving to what).

I took `Output.toDenotation` and expanded that so that every `Expr` has it, so that we can look into an arbitrary expression and see what outputs are in it.  I haven't actually implemented it for every kind of Expr, though, a bunch are still implemented trivially.  But for basic uses (like when the cursors are all near the root of the tree) it gives us the same result as prevOutput used to contain.

This is no longer the actual *denotation* of the Expr, however, so I renamed the function to `getOutput()`

(2) In every tick of the generator, a *lot* more is getting done now -- it's equivalent to the whole "clock" cycle of PriorityExpr.  Within that cycle, it had gotten really hard to actually read what delta/deriv logs were related to one another.  Derivs on different tapes don't happen in a clock-like sequence anymore (with a handy prevOutput after each one as a signpost), it's recursive, so they form a tree.  It's not straightforward to read that tree off the bare, unindented log messages -- possible, but a real pain.

So I put in the ability to increment and decrement an "indentation" parameter in the logging system, letting us view the delta/deriv logs in their tree form.

So long as you have enough width to view it, this ends up being pretty readable and succinct.  Each tree contains what were (in some cases) dozens of generator ticks, but instead of having to scroll back and manually search for the preceding node, it's right there as the parent in the tree.

(Each generator tick is still its own tree, but that's fine, I think it would get too busy and deeply-nested otherwise.)