Experimental implementation of Trees that Grow
This is an experiment in implementing [a simpler version of] "Trees that Grow" (Najd & Jones, 2016): https://www.microsoft.com/en-us/research/uploads/prod/2016/11/trees-that-grow.pdf  (I don't necessarily think we'll do this, but I wanted to commit this for reference.)

We've been wanting to redo how tape information is stored; right now it's an ephemeral piece of mutable state that disappears after every pass.  (That's because it used to change fairly dramatically, but recent changes have settled it down and now we're mostly recalculating the same thing over and over.)  The jankiness of tapes is a continual source of bugs, because it's quite easy to refer to it when it doesn't exist.

That re-raises the question of exactly how to store the tapes, because it's a property of the AST that comes into being at one point and goes away later.  The simplest solution is, of course, "Have the property always exist on all nodes and just don't use it except during passes where you know it's valid".  (Call that Solution 1.) That works, but even though it typechecks, it's not really faithful to the spirit of type safety.

There's also a Solution 2, just bite the bullet and make a whole new set of node types (e.g., when you need a SequenceGrammar with tapes, make a new class called TapedSequenceGrammar that has them).  But do we really want to duplicate the whole hierarchy every time a pass adds or removes information?

So far our compiler goes for a patchwork of Solutions 1 and 2.  We still have a distinction between TSTs, Grammars, and Exprs in part because they have some pass-specific information in them like TSTs have `.pos` and Grammars have `._tapes`.  But even then, we use Solution 1 at the same time, like even though all Grammars have `._tapes` you can't be sure that referring to it is valid -- there are grammar stages before tape inference.

This is actually a general problem, it's bitten us before, and it bites most compiler writers eventually.  You have a bunch of AST nodes and a bunch of passes.  Associated with those nodes are pieces of information that only exist through certain spans of passes and only on certain kinds of nodes (e.g. a "type" property that only exists on expression nodes but not statement or namespace nodes, but also only exists on those nodes in between type inference and type erasure).  To some extent you need a "two-dimensional" type system, where one axis is the kind of node and the other axis is "time".  Think of it like a `(type, type) => type` function.

The other complication is that you need to be able to assert this over the entire tree.  Like TypeScript's intersection types make it easy to say `SequenceGrammar & {tapes: string[]}`, so we should be able to define our pass as being of type `Grammar => Grammar & {tapes: string[]}` right?  Nope, that only asserts that the grammar at the root has `.tapes`, whereas we need to assert that every grammar in the tree that is supposed to have tapes then, does.

There's a third solution, presented in Najd & Jones (2016), which is currently considered the Right Thing to Do here.  Najd & Jones use Haskell's higher-kinded types as a way to effectively get that two-dimensional type system, where the actual properties of a node depend both on the kind of node it is and what stage of compilation we're on.

TypeScript doesn't have higher-kinded types, but it can express effectively arbitrary functions on types.  (TS's type system is itself Turing-complete.)  Yamamoto Yuji figured out how to do Trees that Grow here, effectively taking advantage of the fact that interfaces are themselves (type => type) functions: https://dev.to/igrep/flexiblly-extend-nested-structures-trees-that-grow-in-typescript-4347

In the end, we don't need the full complexity of that Yamamoto's system affords.  I'm perfectly happy adding `.tapes` to every node in between tape inference and erasure, I don't really need to make it dependent on *both* the pass and the node type.  This simplifies things considerably, we really only need a (type => type) function, which we can implement either as an `interface` or `type`, then intersect the result of that function with the node type.

Anyway, I'm not commiting to doing this, but I now know how to do it, and we should remember for the future that there's an alternative to Solution 1 and Solution 2.