New repetition, improved nontriviality detection
(1) This commit expresses repetition in a different way.

The tricky bit about multi-tape repetition, that doesn't occur in single-tape repetition, is that in A*, A might have nullable content on one of its tapes, say t1.  If this were a single tape, we could just simplify the whole thing to epsilon, but for multiple tapes that loses content -- there might be content on t2, say.  (We also don't want to just delta out on t1, because t1 might be `t1:a|eps`, we don't want to simplify that because we'd lose the possibility of `a`.)

Our current solution is procedural in nature -- when the RepeatExpr is in the situation where it'd be choosing that null path, it emits an epsilon.  That doesn't add anything to the output, but it does let us go down the derivational path that will let us get that content on t2.

This solution gets the right answers but has two downsides.  One, we can end up with identical outputs but there's nowhere in the system to eliminate that by disjoining, because they lie on different derivational paths.  The other is just that it's hard to explain exactly what happens, it'd be better to be able to express the answer as a formula.

This commit tries out a NewRepeatExpr that, instead of emitting epsilon, expresses the result analytically:

    D_t:a(X) = d_t(X)* ⋅ D_t:a(X) ⋅ X*

Think about this one like this.  When `t1` is nullable in `X`, and D_t:a(X) succeeds and emits `t1:a`, there are multiple ways to have gotten there.  You could have just emitted `a`, but you also could have (abstractly) emitted an epsilon and then emitted `a`, or two epsilons, or three, or infinite before you emitted `a`.  You need to capture the idea that there might be infinite epsilons you emitted, and capture any content on other tapes that corresponds to those epsilons.  That's what the extra term `d_t(X)*` does.

(2) This doesn't yet work for everything because it runs into the partial threading of Cursors/Outputs through our system.  (Like some examples we disjoin cursors, which is theoretically fine but UnionExpr still has a trivial implementation of `forward()`.)

I left that unfixed for a moment, though, because that configuration revealed a hole in one of our safety nets, the one that detects that we went nowhere.  It's a case where an expression Cursor_t1(X) where t1 is nullable, and X does not simplify, but `handled` returns true, and we just infinitely keep calling `forward()` on it.

So really, it's not just that Cursors should return true for `handled` just because they intercepted the `forward()` call.  A Cursor should return true if either the cursor emitted an actual character, or if its child did.  (That is to say, `handled` is the property `we emitted a character on some tape`.)  Then, we throw an error in the generation loop if:

    * the result is not fully simplified (i.e. not null/epsilon/output), and
    * the forward is unhandled (i.e. no tape emitted an actual character)

This lets us detect that we're churning and going nowhere, but still allows the situation where we're simultaneously delta'ing out of every remaining tape and yielding.