Further development of StarExpr
The NewRepeatExpr had a few snags:

(1) While it's not directly related to repetition, testing the NewRepeatExpr revealed a subtle issue in alternation.

UnionExpr is rare in that it doesn't wrap its children's derivatives in a new UnionExpr, it just yields them.  (After all, the results of yields are themselves semantically in alternation to each other already.)

These sometimes turn into UnionExprs again, though, when caught by `disjoin()`, which will union them up again as appropriate.

This can result in a rearrangement of the scope of alternation, it's possible to have one step be `X(Uni(Y,Z))` and after a round of derivation have that come out as `Uni(X(Y),X(Z))`.  And if `X` is a cursor, this means that the outputs of Y and Z cannot be disjoined.

So it's necessary now for UnionExpr to (like other Exprs) wrap the derivatives of its children in a UnionExpr.

(2) While the NewRepeatExpr worked for min=0, max=Infinity (that is, X*), it didn't actually work for other numbers.

Let's consider a naive implementation for exactly `n` repetitions:

    D_t:a(X^n) = d_t(X)^(n-1) ⋅ D_t:a(X) ⋅ X^(n-1)

This is the wrong number of repetitions, if n=5 we don't want that first term repeating 4 times and the last one also repeating 4 times.  We want there to be a *total* of 4 repeats.  This means that the number of repeats of the last term depends on how many times the first has repeated, it's:

    D_t:a(X^n) = d_t(X)^(k) ⋅ D_t:a(X) ⋅ X^(n-k-1)

We don't actually have a mechanism for two terms to constrain each other in this way.  I suppose we could, but ultimately we don't really use X^n repetitions.  We use them for testing, there's not actually a syntactic means for a Gramble programmer working in sheets to create an X^n.  So for now I'll put that on the back burner and use the old implementation of RepeatExpr, and rename RepeatExpr as StarExpr.