Threading .forward through Embeds
Embeds had a trivial `.forward()` before this, which means that we couldn't have Cursors inside symbols.  We weren't yet using Cursors like this, but we'll want to in the future -- it'll let us do things like true semantic Hiding and collection of intermediate outputs in rule cascades.

A big side benefit of threading .forward through Embeds is it gets rid of the whole complication with simplification of `Embed(ε)` and the `.END` tape hack.  The big issue here is that we're not able to perfectly simplify expressions upon their construction (during `.constructExpr()`), e.g. when `X` is ε, `Embed(X)` is too, but `X` might not be constructed yet.  Only once the entire thing has been constructed, can we simplify it.

Going through a single cycle of `delta/deriv` is sufficient to simplify it, since each of these results in simplified outputs, but something like the `Embed(X), X = ε` grammar above doesn't actually go through a single cycle of `delta/deriv`, because delta and deriv are relativized to tapes and this grammar has no tapes at all.  This forces us to either (a) write a dedicated simplication pass that happens regardless of tapes or (b) run the delta algorithm once with a "dummy" tape.  (a) was tedious, so as a quick hack I made it so epsilons and nulls reported that they had a dummy tape, so that delta would automatically run even on these tapeless grammars.

We don't have to do that anymore, though; `.forward()` can do what we need.  It's called after constructExpr is complete, it's not relativized to any tapes, it's always called at least once, and its output is in simplified form.

Doing this eliminates the need for a symbol table parameter in `constructEmbed` (we're no longer needing to look anything up from that argument, we're using the one in the `env` argument to `.forward()`), and in turn that lets us remove that same parameter from `constructExpr`, and so callers of `constructExpr` no longer have to worry about creating or finding the relevant symbol table.