Threading forward/outputs through UnionExpr
This was actually a tricky one.

If we want to have local cursors (i.e. not just at the top level), then of course we have to allow things like `Cursor(t1, t1:hello) | Cursor(t2, t2:world)`.  However, on the next step, this becomes `Output(t1:hello) | Output(t2:world)`.

This is something new to the system: an alternation of outputs.  This isn't recognizeable to the generator as a valid finish (it's not itself an output, epsilon, or null), and nor is there a way to simplify this into one (Outputs don't admit alternation).  Until now every output corresponded to a single StringDict.

Since we need to be able to express things like this, this means Outputs now have to be able to contain some sort of alternation, and that they correspond to `StringDict[]` instead.

(a) Prior to this commit, OutputExprs were fundamentally concatenations.  They performed the role that was previously the job of ConcatExpr, and prior to that was the job of OutputTrie.  But all of these were the same idea, they all expressed concatenation.

Now, they have to potentially express alternation as well.  So rather than have different kind of OutputExpr for concatenation vs. alternation, OutputExpr is now a UnaryExpr that can contain ConcatExprs and UnionExprs as need be.

(b) `Expr.getOutputs()` now returns a `StringDict[]` rather than a `StringDict`.  `ConcatExpr.getOutputs()` thus has to give the product of its children's outputs, rather than just the concatenation.  (It's the same as when OutputTries could contain BitsetTokens -- same return type, same basic algorithm to get it.)