Handling nested Cursors referencing same tape
Our compiler shouldn't be constructing grammars like `Cursor(t1, Cursor(t1, X))` in the first place, but this is a possible arrangement of nodes, and it's worth pinning down its semantics.  Does it have a denotation?  How should we evaluate it?

Before this commit, if `Cursor(t1, child)` received a deriv/delta on any tape, it just forwarded the query on to its child, even if that tape was `t1`.  With nested `t1` cursors, this results in a scrambled output.  What happens is that each Cursor sends its own `t1` query (as if there were two `t1` ticks in the big Priority clock), and then stores the result in its Output.  This will consume all the `t1` content in the child, but each Cursor will get an alternate character.  (E.g. if the child is `t1:hello`, one will collect "hlo" and the other "el".)  At the end, the resulting Outputs are effectively concatenated, giving `elhlo`.

Even if we decide `Cursor(t1, Cursor(t1, X))` is an erroneous configuration, this can't be its denotation.  (Aside from being crazy, it would result in semantic inconsistency, because `Cursor(t1, X)` would depend on the *form* of `X` rather than its denotation.  You could substitute in two different expressions of the same output for `X` and get different results.)

The easiest (and probably ultimately the correct) semantics for this situation is a "hide" semantics.  `Cursor(t1, X)` hides `t1` outside of itself; a reference to tape named `t1` outside of this node simply refers to some other tape.

Nesting another `Cursor(t1, ...)` around this has no effect, it's the same as wrapping it in a Cursor referring to a non-existent tape.  (Which is fine, by the way.  It's semantically well-defined to ask "What is the deriv/delta with respect to tape `kjaj0923u8409jafds` of the grammar `t1:hello`?"  Its deriv will always be NULL and its delta will be `t1:hello`.  We might decide to issue a warning or something in these situations because such configurations probably derive from an error on the programmer's part, but they don't cause any *semantic* problems for us.)