Small change to "handled" detection
Reminder of background: We want to detect when `.forward` goes nowhere, because that'll result in an infinite loop.  But we don't want to have to calculate if the input and output are equal, because that's expensive.  So we have a flag `handled` instead, it's a surrogate for saying whether the output state is or isn't the same as the input.

A few of our more complex RepeatExpr tests were throwing spurious exceptions (they were moving to a different state, but still coming back as unhandled).  It's because we were treating Epsilon token results from deriv (which come from RepeatExpr) as if they were handled=false, but these actually do result in a different output state, so really they're handled=true.  A small change to the `handled` calculation makes this so.