Threading outputs through rule join machinery
Prior to this commit, you couldn't have Outputs within rule joins, which was inadequate because one of the reasons for a smarter outputting system is so that we can reveal the intermediate forms of phonological rules.

This commit makes some of the rule join machinery aware of Outputs so that they properly simplify rather than throwing exceptions.  The most important of these is probably Join itself.

In the future we may want to look again at what's the "correct" result for an arbitrary join of OutputExprs, but for now the rule is that the output of the left side is the output of the whole thing, unless the left side is epsilon and the right side is an OutputExpr.