Refactored the repeat tests (testGrammarRepeat) as per commit 87740a1.
Also added more * repeat tests. Note: now all tests involving * repeats are so
denoted by a * suffix added to the test number.

Included in the additions are tests exploring issues with nested infinite repeats.
These are tests 38-1* through 38-19*,

Finally, there are a large number of tests for repeats of nullable matches,
many of which are basically redundant. I added the skipGeneration option to a
bunch of these tests to reduce the running of redundant tests. We can decide
later if we just want to remove the skipped tests completely.