Updating the README
Now that we're going public, we need more README than just how to install it.  Now it gives a brief introduction to Gramble.

The screenshot is out-of-date syntactically (it uses ":" for assignment still) but I'll update that soon; for now I just want to make sure we can embed images and see how it looks.