Porting StarExpr from derivUnify branch
The new repetition semantics from the derivUnify branch fixes the problems in nested repetition (and more broadly, repetition of nullables) that we found.

The problem was is that certain expressions -- those that are both nullable and can emit (like eps|t1:a) -- will, when repeated infinitely, have a derivational path that of infinite epsilons.  This isn't wrong exactly -- they are valid paths in the sense that, yes, we might have 1,000,000 epsilons on t1 before emitting "a", and there might be a t2:b that corresponds to this epsilon so we can't NOT go down this path -- but still, we can't allow an infinite loop like this.

In the derivUnify branch, this issue is already solved a different way.

D_t1:a(X*) = (d_t1(X))* . D_t1:a(X) . X*

The first term there expresses in analytic form the results of going down the "epsilon branch" up to an infinite number of times (including, e.g., the result of a potentially-infinite repetition of t2:b that might be associated with an epsilon on t1).

This is now what's created for `X{0,Infinity}` repetitions.  For `X{n>0,Infinity}` repetitions, we construct `X{n,n} . X{0,Infinity}`.

These two changes address all the failing tests in testGrammarRepeat.