Removing RepeatExpr
When separating stars (that is, X{0,Infinity}) and repeats (other combinations), I was wondering if expressing non-finite repetitions as sequences and alternations (with no repetitions) would help with the problem of generating identical grammars.

(This is something we used to do, actually, before we had figured out a better semantics for Repeat Expr.  We ended up taking it out because it made Expr.id hard to read.  But now getting rid of duplicate answers is a higher priority than easy-to-read ids.)

The hypothesis was that, because the cause of most (all?) duplicate answers were epsilon token returns, and one source of those is RepeatExpr, that we could get rid of many duplicate answers by eschewing RepeatExpr and instead constructing an equivalent grammar out of StarExprs, ConcatExprs, and UnionExprs.

This hypothesis seems to be borne out -- many of the spurious duplicates on this branch are eliminated by this change.  (It's possible many of the old duplicates are eliminated too, but I haven't checked.)

Not all of them, however!  I think these may be due to the epsilon token returns from CorrespondExpr.  Returning these is, I think, unavoidable right now.  It may be possible to avoid duplicates, though, by changing the scope at which we process epsilon token returns.