Adjusting a few tests
(1a) Which tests generate duplicate outputs has changed somewhat compared to the main branch.  I've gone through the tests that now fail due to duplicate outputs, and determined that they're valid duplicates (they're due to nontrivially different derivational histories), and put WARN_ONLY flags on them.

(1b) Side note: I didn't do the other way around, going through tests that now DON'T generate duplicates and removing the WARN_ONLY flags.  Those are mostly in a file that's soon to be reworked into the new testGrammar framework anyway, so it'll be less confusing to do that later.

(2) As mentioned in the last commit, the ad-hoc mechanism that prevented `() -> C & ab` generating things like `CCCCCCCCCCCCab` no longer functions due to other changes, and has been removed.  I've changed the relevant test to account for this.  (We might not *want* this behavior and add a different mechanism to prevent it, but it's correct behavior for the current semantics so the test should reflect that.)