Making unit tests use Cursor
There are two places where we "stage" grammars/queries for execution, one in Interpreter and the other in the executeTests pass.  The former had been updated to use Cursor, but the latter still used Priority.  This updates it to use Cursor instead and restores the sheets tests that had been failing due to this.