Cleanup of deprecated code
It's become clear that a more sophisticated (or rather, simpler but more semantically correct) compilation system has become necessary, which will require some refactoring of the tape/vocab system or the complexity will be overwhelming.

I started on this refactoring (you can see the beginnings in a new unfinished `calculateTapes` pass), before realizing that this refactoring is complicated by deprecated and (in actual Sheets) unused code that still enshrine older ways of thinking about how tapes work.  This commit finally removes them.

(1) Remaining references to the dummy tape `.END` system

(2) Remaining references to the Priority system

(3) Same-tape replacement rules, and the JoinReplace system that make them possible (in which programmers can specify arbitrary sets of tapes for replacement rules to act on).  It hasn't actually been possible for a programmer to access this system for quite a long time, but we've been keeping it around and it's complicating reasoning about grammars' tape structure.

Removing JoinReplace also makes two test suites disappear (`testJoinReplace` and `testNotReplace`), although most of those are exploring problems that are no longer "live" concerns.  The tests that are still relevant have already been ported to `testGrammarJoinRule`.